Deve hl

+++ backend/moviedb/src/main/java/com/group12/moviedb/controller/MoviesToWatchController.java
@@ -2,6 +2,8 @@ package com.group12.moviedb.controller;
 
 import java.util.List;
 
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
@@ -10,7 +12,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.group12.moviedb.repository.MovieRepository;
 import com.group12.moviedb.repository.MoviesToWatchRepository;
+import com.group12.moviedb.repository.UserRepository;
 import com.group12.moviedb.models.Movie;
 import com.group12.moviedb.models.MoviesToWatch;
 import com.group12.moviedb.models.User;
@@ -18,60 +22,84 @@ import com.group12.moviedb.models.User;
 @RestController
 public class MoviesToWatchController {
 
-    private final MoviesToWatchRepository moviesToWatchRepository;
+    @Autowired
+    private UserRepository userRepository;
+    @Autowired
+    private MovieRepository movieRepository;
+    @Autowired
+    private MoviesToWatchRepository moviesToWatchRepository;
 
-    public MoviesToWatchController(MoviesToWatchRepository moviesToWatchRepository) {
-        this.moviesToWatchRepository = moviesToWatchRepository;
-    }
-
+    @CrossOrigin(origins = "*")
     @GetMapping("/movies_to_watch")
     public List<MoviesToWatch> getMoviesToWatch() {
         return moviesToWatchRepository.findAll();
     }
 
+    @CrossOrigin(origins = "*")
     @GetMapping("/movies_to_watch/user/{user_id}")
-    public MoviesToWatch getMoviesToWatchById(@PathVariable("user_id") int userId) {
-        User user = new User();
-        return moviesToWatchRepository.findByUser(user).orElse(null);
+    public List<MoviesToWatch> getMoviesWatchedById(@PathVariable("user_id") int userId) {
+        User user = userRepository.findById(userId).orElse(null);
+        if (user == null) {
+            return null;
+        }
+        return moviesToWatchRepository.findByUser(user);
     }
 
+    @CrossOrigin(origins = "*")
     @GetMapping("/movies_to_watch/movie/{movie_id}")
-    public MoviesToWatch getMoviesToWatchByMovieId(@PathVariable("movie_id") int movieId) {
+    public List<MoviesToWatch> getMoviesToWatchByMovieId(@PathVariable("movie_id") int movieId) {
         Movie movie = new Movie();
-        return moviesToWatchRepository.findByMovie(movie).orElse(null);
+        return moviesToWatchRepository.findByMovie(movie);
     }
 
+    @CrossOrigin(origins = "*")
     @GetMapping("/movies_to_watch/user/{user_id}/movie/{movie_id}")
-    public MoviesToWatch getMoviesToWatchByIds(@PathVariable("user_id") int userId, @PathVariable("movie_id") int movieId) {
-        User user = new User();
-        Movie movie = new Movie(); 
-        return moviesToWatchRepository.findByUserAndMovie(user, movie).orElse(null);
+    public List<MoviesToWatch> getMoviesToWatchByIds(@PathVariable("user_id") int userId, @PathVariable("movie_id") int movieId) {
+        User user = userRepository.findById(userId).orElse(null);
+        Movie movie = movieRepository.findById(movieId);
+        if (user == null || movie == null) {
+            return null;
+        }
+        return moviesToWatchRepository.findByUserAndMovie(user, movie);
     }
 
+    @CrossOrigin(origins = "*")
     @PostMapping("/movies_to_watch")
     public MoviesToWatch addMoviesToWatch(@RequestBody MoviesToWatch moviesToWatch) {
         return moviesToWatchRepository.save(moviesToWatch);
     }
 
+    @CrossOrigin(origins = "*")
     @PostMapping("/movies_to_watch/user/{user_id}")
     public MoviesToWatch addMoviesToWatchById(@PathVariable("user_id") int userId, @RequestBody MoviesToWatch moviesToWatch) {
-        User user = new User();
+        User user = userRepository.findById(userId).orElse(null);
+        if (user == null) {
+            return null;
+        }
         moviesToWatch.setUser(user);
         return moviesToWatchRepository.save(moviesToWatch);
     }
 
+    @CrossOrigin(origins = "*")
     @PostMapping("/movies_to_watch/user/{user_id}&movie/{movie_id}")
     public MoviesToWatch addOrUpdateMoviesToWatch(@PathVariable("user_id") int userId, @PathVariable("movie_id") int movieId, @RequestBody MoviesToWatch moviesToWatch) {
-        User user = new User(); 
-        Movie movie = new Movie(); 
+        User user = userRepository.findById(userId).orElse(null);
+        Movie movie = movieRepository.findById(movieId);
+        if (user == null || movie == null) {
+            return null;
+        }
         moviesToWatch.setUser(user);
         moviesToWatch.setMovie(movie);
         return moviesToWatchRepository.save(moviesToWatch);
     }
+    
     @PatchMapping("/movies_to_watch/user/{user_id}")
     public MoviesToWatch updateMoviesToWatchByUserId(@PathVariable("user_id") int userId,
                                                      @RequestBody MoviesToWatch moviesToWatch) {
-        User user = new User();
+        User user = userRepository.findById(userId).orElse(null);
+        if (user == null) {
+            return null;
+        }
         moviesToWatch.setUser(user);
         return moviesToWatchRepository.save(moviesToWatch);
     }
@@ -80,8 +108,11 @@ public class MoviesToWatchController {
     public MoviesToWatch updateMoviesToWatchByIds(@PathVariable("user_id") int userId,
                                                    @PathVariable("movie_id") int movieId,
                                                    @RequestBody MoviesToWatch moviesToWatch) {
-        User user = new User();
-        Movie movie = new Movie(); 
+        User user = userRepository.findById(userId).orElse(null);
+        Movie movie = movieRepository.findById(movieId);
+        if (user == null || movie == null) {
+            return null;
+        }
         moviesToWatch.setUser(user);
         moviesToWatch.setMovie(movie);
         return moviesToWatchRepository.save(moviesToWatch);
@@ -92,8 +123,11 @@ public class MoviesToWatchController {
                                                    @PathVariable("movie_id") int movieId,
                                                    @PathVariable("note") String note,
                                                    @RequestBody MoviesToWatch moviesToWatch) {
-        User user = new User(); 
-        Movie movie = new Movie(); 
+        User user = userRepository.findById(userId).orElse(null);
+        Movie movie = movieRepository.findById(movieId);
+        if (user == null || movie == null) {
+            return null;
+        }
         moviesToWatch.setUser(user);
         moviesToWatch.setMovie(movie);
         moviesToWatch.setNote(note);
@@ -102,16 +136,21 @@ public class MoviesToWatchController {
     
     @DeleteMapping("/movies_to_watch/user/{user_id}")
     public void deleteMoviesToWatchByUserId(@PathVariable("user_id") int userId) {
-        User user = new User(); 
+        User user = userRepository.findById(userId).orElse(null);
+        if (user == null) {
+            return;
+        }
         moviesToWatchRepository.deleteByUser(user);
     }
     
     @DeleteMapping("/movies_to_watch/user/{user_id}/movie/{movie_id}")
     public void deleteMoviesToWatchByIds(@PathVariable("user_id") int userId,
                                           @PathVariable("movie_id") int movieId) {
-        User user = new User();
-        Movie movie = new Movie(); 
+        User user = userRepository.findById(userId).orElse(null);
+        Movie movie = movieRepository.findById(movieId);
+        if (user == null || movie == null) {
+            return;
+        }
         moviesToWatchRepository.deleteByUserAndMovie(user, movie);
     }
 }
-
